03_Lang-chain/requirements.txt

langchain
langchain-community
langchain-openai
openai
python-dotenv
pypdf
langchain-core
langchain-chroma
langchain-google-genai
